+++ docker-compose.yml
@@ -1,0 +1,11 @@
+version: '3.9'
+
+services:
+  postgres:
+    image: postgres:latest
+    env_file:
+      - .env
+    ports:
+      - '5432:5432'
+    volumes:
+      - ./data/postgres:/data/postgres
