+++ frontend/public/assets/bective-logo.png
