+++ frontend/src/styles/NavBar.scss
